Moved Kryo extension functions into Kryo.kt
